step2 2차 코드 리뷰 반영
