+++ build.gradle
@@ -13,6 +13,5 @@ dependencies {
     compile('com.sparkjava:spark-template-handlebars:2.7.1')
     compile('ch.qos.logback:logback-classic:1.2.3')
     testCompile('junit:junit:4.12')
-    testCompile('org.mockito:mockito-all:1.10.19')
     testCompile('org.assertj:assertj-core:3.11.1')
 }
