+++ src/main/java/racinggame/RacingGame.java
@@ -4,12 +4,9 @@ import racinggame.domain.Car;
 import racinggame.domain.RacingResult;
 import racinggame.service.CarAdvanceService;
 import racinggame.service.RandomValueGeneratorImpl;
-import racinggame.view.InputView;
-import racinggame.view.ResultView;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Scanner;
 
 public class RacingGame {
 
@@ -21,25 +18,7 @@ public class RacingGame {
         this.configuration = configuration;
     }
 
-    public static void main(String[] args) {
-        InputView inputView = new InputView(new Scanner(System.in));
-        String[] carNames = inputView.getCarNames();
-        Integer numberOfTries = inputView.getNumberOfTries();
-
-        RacingGame racingGame = new RacingGame(new Configuration(carNames, numberOfTries));
-
-        ResultView resultView = new ResultView();
-        resultView.showResultTitle();
-
-        RacingResult result = new RacingResult(racingGame.createCars());
-        for( int i = 0; i < numberOfTries; ++i ) {
-            racingGame.proceed(result);
-            resultView.showCarPositions(result);
-        }
-        resultView.showWinners(result);
-    }
-
-    private List<Car> createCars() {
+    public List<Car> createCars() {
         List<Car> cars = new ArrayList<>();
 
         for( String name : configuration.getCarNames() ) {
@@ -49,7 +28,7 @@ public class RacingGame {
         return cars;
     }
 
-    private void proceed(RacingResult result) {
+    public void proceed(RacingResult result) {
         carAdvanceService.moveForward(result.getCars());
     }
 }
